+++ extensions/positron-python/src/test/positron/extension.unit.test.ts
@@ -10,9 +10,9 @@ import { SemVer } from 'semver';
 import { IInstaller, InstallerResponse } from '../../client/common/types';
 import { IInterpreterService } from '../../client/interpreter/contracts';
 import { IServiceContainer } from '../../client/ioc/types';
-import { checkAndInstallPython } from '../../client/positron/extension';
 import { Architecture } from '../../client/common/utils/platform';
 import { EnvironmentType } from '../../client/pythonEnvironments/info';
+import { checkAndInstallPython } from '../../client/positron/manager';
 
 suite('Set up extension', () => {
     let serviceContainer: TypeMoq.IMock<IServiceContainer>;
